feat: send email on resource OrderAdressResource

+++ app/models/order_model.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 
 from sqlalchemy.orm import relationship, backref
 from sqlalchemy.sql.schema import ForeignKey
-from sqlalchemy import Column, String, Integer, TIMESTAMP, Boolean
+from sqlalchemy import Column, String, Integer, TIMESTAMP, Boolean, Float
 
 
 @dataclass
@@ -13,7 +13,7 @@ class OrderModel(db.Model):
     id: int
     invoice_url: str
     was_paid: bool
-    total_price: int
+    total_price: float
     created_at: datetime
     payment_type: str
     payment_day: datetime
@@ -23,7 +23,7 @@ class OrderModel(db.Model):
     id = Column(Integer, primary_key=True)
     invoice_url = Column(String)
     was_paid = Column(Boolean, default=False, nullable=False)
-    total_price = Column(Integer, nullable=False)
+    total_price = Column(Float, nullable=False)
     created_at = Column(TIMESTAMP, default=datetime.now())
     customer_id = Column(Integer, ForeignKey("customers.id"))
     address_id = Column(Integer, ForeignKey("addresses.id"))
